SnabWall: Add end spy-pcaps test runner
This adds a selftest.sh script for SnabbWall, with one first test suite which
exercises the "snabb wall spy" command using capture files as inputs. For each
capture file, a corresponding .spy.expect file must contain a subset of the L7
protocols expected to be detected by "snabb wall spy", one per line. If any of
the expected L7 protocols is not detected from the capture, the test case is
considered to be failed.

One first test case with a capture file from https://wiki.wireshark.org/SampleCaptures
is provided as well.