Do not leak packets in apps.wall.util.SouthAndNorth
The SouthAndNorth can be used as a kitchen-sink if no outputs are connected.
This means that is must handle deallocating packets that it is not pushing
to the next application in the chain. The changes in this commit also handle
gracefully the case in which one of the :on_{north,south}bound_packet(p) may
return a different packet: the "p" packet is freed and the one returned by the
callback sent instead.