Streamline SNMP config in l2vpn
With the new counter framework in interface drivers and the
interface-specific SNMP code factored out to
lib/ipc/shmem/iftable_mib.lua it is possible to move the SNMP
configuration from the interface-level to the global config level and
add SNMP for sub-interfaces.

The README has been updated to reflect the new mode of configuration.