Hardened IPv6 reassembly
Use a fixed amount of memory for reassembling IPv6 fragments, on a ctable.
It only allocates once, on initialization, and never frees memory.
When the backing table is full, it randomly ejects fragments.

Tests include out-of-order reassembly.
Bounds checking is done to avoid buffer overflows, regardles of input offsets.

This includes a to-memory clone for Igalia's struct packet implementation.
This allows IPv6 reassembly to avoid a subtle packet corruption bug on
implementations where packet.data is an abstraction rather than a raw buffer.