virtio net_device: JIT each connection separately
The vhost_user app now JITs separate machine code for each connection.
The transmit and receive paths are also JITed separately.

This improves performance and consistency for many workloads,
especially for dealing with different combinations of virtio-net
options and with virtual machines that switch device drivers (e.g.
from Linux to DPDK or Snabb).

The root problem was that our vring processing code is written in a
style that performs badly with LuaJIT. The code is implemented in one
central loop, VirtioVirtq:get_buffers(), which has many different
behaviors depending on its arguments (callback functions, negotiated
virtio features, etc). LuaJIT compiles all of these different options
into a growing network of "side traces" and this degrades performance.

To make the code perform efficiently we need to make its control flow
more consistent from the point of view of the JIT.

One solution to this problem would be to write specialized versions
of the vring processing code for all different combinations of situations:

- Separately for transmit and receive;
- Separately for mergeable RX buffers and without;
- Separately for indirect decsriptors and without;
- etc...

However doing this by hand would be a lot of work. We would have to
rewrite the vring processing code and the new version would be much
more code.

Instead we make the JIT do this work for us automatically. We just
have one copy of the source code but we load a fresh copy of the
object code for each vring. This means that each vring is JITed
separately in a way that suits its specific behavior. So the machine
code for a vring that supports mergeable RX buffers will automatically
optimize for that case, and so on.

You can think of it as creating many different vring processing loops
that each inline a different set of subroutines.

If you want to know more about doing this kind of optimization with
LuaJIT then there is some background information here:

  https://github.com/LuaJIT/LuaJIT/issues/208#issuecomment-236423732

Note that the callback-driven nature of the vring processing code is
not a problem directly. Callback indirection compiles very efficiently
when the same callback function is used every time. However,
performance degrades when the JIT is sharing machine code between
multiple calls that provide different callback functions. (The same
thing happens if you pass different parameter values that cause 'if'
statements to switch from 'then' to 'else'.)

Quoth the LuaJIT masters all too innocently:

> Avoid unbiased branches.