If there is a br address for a bad softwire, use it
The lwAFTR sends ICMPv6 unreachable (code 5) messages if it gets a message
from a bad softwire (ie, from a B4 that is not in the binding table, or to
a br address that is not associated with that B4). If a br address can be
associated with the IPv6 address being replied to, use that as the source
address; otherwise, use the IPv6 address it was sent to.

In no case should it use the default IPv6 address for the lwAFTR from its
conf file; this is now an obsolete concept.

Added a changelog entry for ICMPv6 unreachable source address fix.