Make the core lwAFTR drop fragments
The lwAFTR was designed with the idea that fragmentation apps around it
would always remove all fragments, and it would never see any. SnabbVMX
allows creation of an app network without the fragmentation apps.
This patch modifies the core lwAFTR app to check for fragments. If it finds
any, it is running in a configuration where fragmentation should be disabled,
so drops them.

This adds and documents two new counters:
drop-ipv4-frag-disabled and drop-ipv6-frag-disabled

There are no new tests that exercise these code paths, yet.