program/worker: Added "worker" program
This is a work in progress. The new command is:

  snabb worker NAME PARENTPID CORE

and the effect is to spawn a new Snabb process that simply executes an
app network and reacts to configuration changes.

The worker process looks in a "group/" shm folder of the PARENTPID to
keep track of its current configuration.

This is intended as an internal interface that the engine can use to
parallelize an app network by delegating parts to different child
processes. The processes would be managed directly by the engine and end
users need not be aware of the "snabb worker" interface.