core.memory: Manage hugetlb file-backing lifetime
HugeTLB memory allocated for DMA is now kept accessible in file-backed
shared memory. This makes it possible for process to map each others'
memory. The new function shutdown(pid) removes file-backing for a
process that is terminating (in order to free memory for the OS).

More specifically:

Allocating a hugetlb page with address xxxxx creates:
  /var/run/snabb/hugetlb/xxxxx.dma         File backed shared memory
  /var/run/snabb/$pid/group/dma/xxxxx.dma  Symlink to above

and the shutdown(pid) function will follow the symlinks to remove the
file backing. (The reason for symlinks is that these directories are on
separate filesystems, one hugetlbfs and one tmpfs.)