hardware.pci: Disable bus mastering on shutdown
The new function shutdown(pid) will disable PCI bus mastering (i.e. DMA)
for all PCI devices that were in use by the process (and its process
group).

This is intended as a safety mechanism to prevent "dangling" DMA
requests on a device after a Snabb process terminates and the DMA memory
is returned to the kernel.

To keep track of which devices are in use we now create shm files with
names like "group/dma/pci/01:00.0" for each device that we have enabled
bus mastering for. These are the devices that are reverted by shutdown().