core.memory: Auto-mapping of DMA memory w/ SIGSEGV
Added a SIGSEGV handler that attempts to automatically map addresses
corresponding to DMA memory. The mapping succeeds when the DMA memory
was allocated by any Snabb process in the process group. If the mapping
does not succeed then the standard behavior of SIGSEGV is triggered.

Includes a simple unit test in core.memory to allocate DMA memory, unmap
it, and then access it anyway. Confirms that the signal handler runs the
expected number of times.