main: Add SNABB_PROGRAM_LUACODE env var
The new environment variable SNABB_PROGRAM_LUACODE can be used to force
a new Snabb process to execute a string of Lua code instead of running
the standard program-selection logic.

This is intended to support Snabb processes that want to spin off helper
processes that need to execute specific code and be independent of the
usual command-line syntax (e.g. sensitivity to argv[0] etc).