snabbnfv traffic: Only re-start engine when configuration changes
Before, snabbnfv would run the engine for one-second intervals, checking
for restart every second and restarting the engine.  However this
interacts poorly with the latency-tracking mechanism, which expects to
be able to call non-performant functions like ffi.typeof() at setup,
because it's just setup.  With this fix we have less trace creation and
interpreter bailout at run-time.