YANG data grammar separates keys, values for tables
A "table" production was previously represented as having a flat value
like a struct, with some of the members marked as keys.  Instead
separate the key from the value, in anticipation of making better
associative maps, and instead of providing "lookup" on the data, provide
"get_entry", "get_key", and "get_value", as ctable does.