Refactor top.select_snabb_instance and fix small issues in claim_name
This fixes a number of issues, some of which are cosmetic. The change
refactors some of the "top" program's code for PID locating and
selection. This change also cleans up a number of checks in claim_name
into asserts.

The big change is removal of the top level import of "core.app" into
main. There doesn't seem to be usage of this paticular import in main
and it causes stale data in "/var/run/snabb" as it creates "engine" shm
directories for the supervisor.