connectx4: IO app now has only 1xSQ and 1xRQ
Previously a single IO app instances could be supported with multiple
send and receive queues. The intention was to share traffic amongst
the queues in order to overcome NIC hardware limitations on per-queue
packet rate.

The problem with that approach is that splitting traffic across queues
potentially changes the relative ordering of packets and this
potentially creates problems for the user. Better to avoid "DWIM" and
have the user arrange multiple IO apps in an application-appropriate
way if indeed they do need to overcome the ~16Mpps per-queue limit on
the ConnectX-4 (100G).