Convert node displaying into data driven code
This converts a if-elseif-else block which was being used
as a mechanism for selecting between different node types
and instead converting it to a more function driven method.

Now there is a table of methods built up that handle the different
types and those are selected based of the node type to handle the
display of the data.