Fix snabbvmx selftest
The packet capture side of things was not getting hooked up correctly
because of a confusion about the names input and output (is it with
respect to the graph, the linke, or each individual app?).  This led to
a PcapWriter that would never write because it never had incoming data.
This went undetected before because such an app would never run, but now
all apps with pull functions run.