Configuration refactor
This PR refactors (re)configuration so that compute_config_actions
results in an ordered list of actions, O(n) in the size of the change
rather than of the graph, such that apply_config_actions can just mindly
apply them.

apply_config_actions is now a mutating procedure which mutates the app
table, the link table, and the current configuration to reflect the
result of applying the config diff in compute_config_actions.