Add Hz parameter to config leader
This lowers the rate at which we call into the kernel to accept new
connections, allowing us to limit the overhead of the config leader.
In the very short term we want to test the config leader in the same
process as the data plane, which will have an associated overhead that
we need to minimize.