core.worker: Updated API & removed dead code
Updated the comment describing the API. (Should be replaced by
documentation in src/README.md.)

Removed config.configure(name, c) which is a remnant of the past design
when we supposed that app network configurations would be lingua franca
for parents to configure their children. That is now out of scope.