Fix perf hazard in leader socket accept()
Although the leader does not have real-time requirements, in this
meantime in which we don't have a multiprocess world yet we are running
it in the data plane, so we need to ensure that it doesn't cause trace
aborts.  Creating a new sockaddr each time was causing trace aborts, so
here we fix it.  Anyway improving this will make traces cleaner.