Refactor lib.yang.xpath.convert_path and fix nits
This completely refactors the convert_path function responsible for taking
a yang xpath path and converting it to a lua path that can be used to resolve
a node in a yang data tree. The refactor in general makes the code more readable
by splitting some of the functionality out of the convert_path function into smaller,
better defined functions.

This also changes the return value from the resolve function to be a function which
takes a data tree and resolves for that data tree. This allows the resolver to be
re-used on multiple data trees with the same schema and path.