Allow ability to parse/print non-struct data
Before, only structs were parseable and printable at the top level.
However considering that we're going to be working with sub-data via the
path API, we need to support parsing and printing sub-data.

Structs print as before.  Tables and arrays print as if they were in a
struct with just that one keyword; that is, as a sequence of KEYWORD {
... } or KEYWORD VALUE; lines respectively.  Sequences turn out to only
be a top-level production so we can remove them from the sub-data
grammar.  The interesting thing comes in with scalars; we're now able to
parse/print them without their keyword so that when you get a
/foo/bar/baz and baz is a scalar, it prints just the value and not the
"KEYWORD VALUE;" ceremony.