Precompute path resolvers before running
Before, we would look at the grammar when resolving a path in data, to
determine what to do at the next path item.  Now we resolve what to do
ahead of time, and just return a closure that does the lookup.

Resolve also returns the grammar for the data that will be looked up, as
a second value.