core.lib: Fix timer() to match documentation
The timer() function now sets its new deadline based on the current time
rather than the previous deadline. This is more consistent with the
documentation and prevents timeouts from being "saved up" in the style
of a token bucket when the timer is not called for a period of time.