Support named lwaftr instances and update ps program
This allows one to specify the name of a lwaftr from the yang configuration
which claims the name in the lwaftr.

It also adds support for this in the new `ps` command. It will now list
the name (preferentially) or if none exists, the PID. It also introduces
a flag (`-p` or `--pid`) which allows for overriding the default behavour of
`ps`. The behavour when using the `-p` flag will only show the PID instead
of the name.