Restart apps when configuration changes in place
Sometimes you might want to update a configuration in place, then
update your app graph to match the configuration.  However Snabb's core
logic to determine which apps need restart after a configuration change
uses `lib.equal` to compare the old app initarg and the new app initarg,
and that strategy doesn't work when we actually end up updating the old
app initarg in place.

To solve this, we build a set of embedded objects that might be updated
in place, for each app in the app graph.  Then, given a change to a
configuration at a path, we determine which objects in that
configuration might be updated in place.  Any app referencing one of
those maybe-updated-in-place objects will need a restart.

There is also a shortcut for specific schemas.  In this way if you know
that, say, adding an entry to the binding table of a lwaftr won't cause
other apps to need to restart, then in that case the schema support code
can avoid needlessly restarting apps.

Fixes https://github.com/Igalia/snabb/issues/602.