container YANG schema nodes always appear in data
It used to be that the Snabb YANG support would expect that "container"
schema nodes would only have corresponding data nodes if they had a
"presence true;" attribute.  This was a misunderstanding of how YANG
expects to map schema to data.  In particular this is a legal schema:

  container foo { leaf bar { type string; } }
  container baz { leaf bar { type string; } }

That is to say, it's perfectly fine for there to be two leaves with the
same name; they don't conflict because the corresponding data would look
like:

  foo { bar "hey"; }
  baz { bar "ho"; }

Before in Snabb we would expect data that looks like:

  bar "hey";
  bar "ho";

And obviously that doesn't work, as you can't tell the leaves apart.

This change adds a "--from" option to "snabb lwaftr
migrate-configuration", indicating the version from which to migrate.
The default is still "legacy", as before, but now there is a new
version, --from=3.0.1, to migrate from the previous mapping between
schema and data.

This change updates all configurations in the source, removes "presence
true" from example and test schemas where it's not necessary, and also
changes to put the binding-table node inside softwire-config.