Fix fast-path bugs in ietf-softwire translator pre_update function
We were never hitting the fast paths because the verb and path arguments
were swapped.  Also we were never detecting that a br-address was
already in the set, because of slight differences between reference and
pointer types (!) -- the br-addresses from the array were reference
types, while the br-addresses from the update were pointers.