Fix killer bug in leader: write beyond array end
Thanks to Asumu Takikawa for tracking this one down.  Quoth
http://luajit.org/ext_ffi_semantics.html in "Initializers":

  Byte arrays may also be initialized with a Lua string. This copies the
  whole string plus a terminating zero-byte. The copy stops early only
  if the array has a known, fixed size.

I was thinking that ffi.new('uint8_t[?]', #str, str) would make an array
with "known, fixed size". But I think that characteristic applies to the
type, and is uint8_t[?] fixed, even when given its length? I thought
so. I was wrong; you can't do this:

  local buf = ffi.typeof('uint8_t[?]', #str)(str)

But you can do this:

  local buf = ffi.typeof('uint8_t[$]', #str)(str)

Using the $ makes an array type with fixed size. The [?] makes a
variable-length array (VLA).

So this bug was that LuaJIT was writing a zero byte beyond the array,
and that was corrupting the GC metadata (probably the header for the
next object).