Skip hw ping to lwAFTR test in CI
The program/lwaftr/tests/hw/test_ping_on_a_stick.sh test cannot be run
inside a container for several reasons:

- The test contains commands, such as iproute2 or lshw,  not supported
in the default image.
- The test binds a NIC to the Linux kernel by PCI address. This enables
the NIC interface. However, when run inside the container, the card gets
rebound to the kernel, but the interface doesn't appear inside the
container but the host. Without an addressable kernel-managed interface,
it's not possible to run the test.