#683 - Generalize migration and fix for printing
The previous migration changes assumed br-address would not be mandatory
meaning that printing would fail. This also wouldn't work for migration
from anything prior to configs at 3.0.1bis. This fixes this in a general
fashion so that any pre-(snabb-softwire-v2) configs can be migrated with
the same code and be parsable with the vanila snabb-softwire-v2 parser.