Allow pfmatch rules to use the flow packet count
This compiles a pfmatch expression for each flow that
matches a pfmatch rule in the firewall rules. Since
there are usually several packets in a flow, this amortizes
the compilation cost across the packets.

This is ok if flows typically have many packets, but
that's not always the case. Unfortunately, it's not
straightforward to compile them ahead of time when using
the $flow_count variable without invasive changes to
pflua.

It's possible to change it so that it only compiles on
each flow if the $flow_count variable is used.