#737 - Support name by --name or snabb-softwire-v1
This adds the ability to define the lwAFTR's name in the YANG lwAFTR
config or the current method of the `--name` parameter on the command
line. The order of presidence is:

1. Use command line argument first (i.e. --name mylwaftr)
2. Use the config parameter at softwire-config.name
3. Don't set a name if no name has been provided.

NB: This will fail if you try to use the same name twice, they must be
unique. This may happen if a name is defined in a config file and it's
not overriden for subsequent instances.