#737 - Add rename program support and fixes bug in claim_name
This adds a function to provide provide a current lwaftr's name and
have it change that name. This involves changing the symlinks on the
file system that provide a name and link to the instance. The change
also changes the code so that it doesn't rely on it being called and
managed in the same process.