#737 - Add support for changing name via snabb set
This adds support for the name to be set (for the first time) or
altered by using the `snabb config set` command on the
`/softwire-config/name` leaf. This will claim the new name and if
there was one release back (free) the old name so it could be used by
another snabb process.

An example of a name change command could be:

   $ snabb config set oldname "/softwire-config/name" "newname"