#737 - Move name claiming to setup_fn and add unclaim_name
This changes where name claiming occurs. The `claim_name` call now
occurs in the `setup_fn` of the lwaftr. This is also how the leader
changes the name of a running lwaftr instance.

The `claim_name` has been reachitected to deal with name changes as
well as successive calls with the same name to allow it to be used how
it is now in `setup_fn`. This also adds the `unclaim_name` allowing
for release or "free"ing of a claimed name. Successive calls to
`claim_name` will now claim the new name and if successful it will
then release/unclaim the previous name.