Change configuration.name and improve name changes
This changes the once `configuration.name` attribute to be it's own
variable `program_name`. This is because `configuration` is an app
graph and it seems wrong to pollute an app graph with the name.

It also improves claiming the name within the lwAFTR. This is done by
having a `switch_name` function which takes an app graph, tries to
claim the name (if one exists) and if it fails, gracefully restore the
old name back onto the app graph's configuration.

It also cleans up some formatting.