Fix #748 - Move claim_name to lwaftr_app
This fixes the regression reported in #748. The regression would mean
that the lwAFTR didn't specify a name when one was given without the
reconfigurable flag enabled. The fix was to move the call to
`claim_name` to `lwaftr_app` which is called both when
`--reconfigurable` is given and without it.