COPYING: Switch to Apache License 2.0 for new code
Notes that LuaJIT is copyright Mike Pall and licensed under MIT.

Notes that new code independently contributed to RaptorJIT is under
Apache License 2.0.

Excellent benefits of the Apache license are the "Grant of Patent
License" and the "Submission of Contributors" that clearly states that
contributors will be covered by the same license as the project
itself.