Remove DUALNUM mode
DUALNUM mode is a complex optional feature, disabled by default on
x86-64, that allows numbers to be stored on the heap either as doubles
(the usual) or as integers (the extension.)

This requires a _LOT_ of ifdef-style conditional logic throughout the
VM and runtime system. My tests have never found a verifiable benefit
to the dual number mode and since it is already disabled by default on
x86-64 I have simply removed it.

This change is a bit scary because it is large and I have mostly made
it manually. The conditional uses of DUALNUM were not ones that I
could remove automatically with 'unifdef'. I have surely made a
mistake in here somewhere and that will surely bit somebody on the
backside one day... Sorry in advance, but meanwhile, onward!