Import revamped nix benchmark framework
The nix code for running benchmarks is now imported in-tree. This
makes it easy to automatically run benchmarks and also generate the
fancy R visualizations. See README.md for example usage.

The visualizations are now generated as PNG files instead of an
Rmarkdown document. This is to make it easier to introduce new
visualizations, or variations of existing ones, and easier to load
large images for zooming.

Under the hood the nix code is also updated so that each test run is a
separate derivation. This means that nix is able to parallelize the
test runs (one execution of each benchmark with one raptorjit version)
and distribute them between machines. This should make the tests run
faster on the Hydra CI cluster and also avoid tying up whole servers
with hours-long derivations that run hundreds of test runs at the same
time. (It also allows you to parallelize test runs on the local machine
to use multiple cores.)