Lookup psid entries using rangemap for range support
This switches to maintaining an instance of the lwaftr's binding table
in the leader to allow for looking up PSID entries, specifically those
which have a range (i.e. specify an "end-addr"). It still needs to
maintain the binding table where possible to prevent it being re-build
on every validation (expensive).