Add snabb set test and fix get_binding_table
This adds another test which checks that erroring occurs on set also
not just add (when setting an invlid softwire). It fixes where the
initial check of the config happens to prevent it being called too
often as well as fixing how often the binding table in the leader
is rebuilt.