engine: Only use class.name if it's a string
The engine has a heuristic for determining the name of an app's class,
for using in profiling reports. One of the rules is to use the 'name'
field of the class if this exists. This is fragile though: if the class
object should happen to have a method called 'name' then we see a type
error i.e. class.name is a function and not a string. (This happens with
subClass() because the parent class does have a name() method.)

Resolved by adding a check to only use class.name if it's actually a string.