Add migration for psidmap on softwire
This adds a minor version for 3.2.0 (3.2.0.1) which moves the psidmap
onto the softwire. This should migrate all configurations away from
using the psidmap block and remove the "end-addr" parameter for good.

The migration can be run on a file using:

    snabb lwaftr migrate-configuration -f 3.2.0.1 <config file path>

This presumes the configuration file is already at 3.2.0 (with the
br-address leaf-list removed).