Convert v2 migration to handle 3.2.0 and 3.2.0.1
This converts the migration added for 3.2.0.1 to be part of the 3.2.0
snabb-softwire-v2 migration. This handles both adding "br-address" on
softwire (away from leaf-list) and psid-map to softwire.

This will not convert something converted "part way" with a previous
version of the migration. If you've already converted a config away from
the br-address list but not yet to the integrated psid-map on softwire
then you'll need to checkout a prior commit to this one and use the
3.2.0.1 converter.

This also removes the bump version which was only used in development.