Update revision statement to include new port-set
This updates the revision statement to bump the date to when the latest
change occured. It also includes the information about the switch from
using the "psid-map" list over to using the "port-set" container on the
softwire itself. This has been done in the same revision as the
br-address as there has been no release with these changes inplace.