Support new softwire in support and binding table
This adds support in the snabb-softwire-v2 support for the new format
where it formats new softwires as a blob the binding table can
quickly add. It also adds a transformer for initial loads of a config
to the binding table so it's able to consume configs in the new
format.

NB: This does not yet handle adding a softwire with a PSID which is
not yet known to data plane (i.e. does not appear on another
softwire).