Change binding table's psid map to ctable
The new snabb-softwire-v2 config no longer allows for ipv4 address
ranges. This means there is no real reason to continue using the
rangemap system and it can be switched to a plain ctable.

This will give us the ability to add port-set entries on the fly
without restarting the lwaftr (how it was previously done) or
rebuilding the entire rangemap (impractical).