Add ability to add psid entries and update tests
This adds the ability to add and remove psidmaps without restarting the
lwaftr. This allows for the addition of softwires with a new psidmapping
without requiring any additional restart. It also updates the unit tests
for the binding table to reflect the new configuration and some of the
internal changes to it.

There still remains quite a bit to do, this does NOT address:

- The leader's version of the binding table isn't updated
- No longer used versions of the PSID map are not removed
- Validation doesn't occur to check the port-set entries line up with
other port-set entries (what should even happen if they don't?)
- Translation layer isn't updated
- Some tests are failing which needs addressing.