Fix a test to use snabb-softwire-v2 schema
This fixes an `snabb config add` test to use the updated schema. This
means adding the psid-map on the softwire using the new `port-set`
container. Previously it was erroring as it was using the old config.

This also removes the no longer valid test to check adding a softwire
without a `psid-map` entry was invalid as entires are now on the
softwire. Whilst this reduces the number of tests I think there will be
additional tests in the (near) future when the behavour around adding
softwires with different port-set info to another softwire with the same
address is decided.