Fix migration to handle larger files and add padding leaf
This fixes the migration so it uses less memory, allowing for larger
files to be migrated. It's currently been tested with a 100MB file.

This also adds back the padding leaf as it causes issues in the ctable
streaming lookup of the softwires. The entry must confirm to certain
sizes or multi_copy.dasl used by ctable streaming will assert an error.