Convert the binding table to use rangemap
This converts the binding table to use rangemap and reintroduces the
padding parameter to snabb-softwire-v2. The rangemap was reintroduced
over the ctable due to the performance impact, as rangemap uses binary
search to find the port-sets, it's notably faster.

The padding parameter was re-introduced as multi_copy.dasl complains
if the size of a ctable (in this case softwire) is not divisable by 4
when multiplied by the entries_per_lookup (max_displacement + 1).