Fix small documentation and other nits
This fixes some of the documentation surrounding the configuration
change. This includes both user facing documentation as well as
commends throughout the code.

It also removes the newly introduced "ignore_duplicates" flag in the
RangeMapBuilder in favour of building a ctable of the added entries in
order to verify that the softwires with the same port-set definition
don't get added multiple times to the psid-map and that they don't
define different values for the port-set.

It removes two now unused functions which were to add or remove psid
map entries. These were removed as the switch back to rangemap makes
it unfeasible to modify the psid-map in the binding table once it's
been built.