Fix structural typing of YANG data
YANG data should be structurally typed.  If I have a struct with just a
"uint16_t a" member, then any other struct that has that same shape
should have the same FFI type.  To implement this we have to memoize the
string -> type function so that all types with the same string
representation use the same type object.  However this is a bit tricky,
givent that types can contain members with compound types -- before,
these nested members with compound types were being treated as nominal
fresh types instead of structural types.  This commit fixes that bug.