Add test for lwaftr's generate-binding-table subcommand
This adds a test which runs the generate-binding-table command and
verifies that it exits with a 0 status code and produces the expected
output for the command line parameters.

It doesn't check the contents of the block but it does check certain
expected things which should verify some problems if they occur.