Fix small performance nit
This fixes a small performance nit in binding_table.lua which would
have been caused by allocating the key of the psid-map on every
lookup. This now reverts it back to how it was, creating the key only
when the binding table is initiated.

This should improve the performance of the BindingTable:lookup
interface a little.