Take default version from ".version" file
The version for the snabb binary will be first taken from the VERSION
environment variable or make argument, then from .version if that's not
set.  The user can also pass EXTRAVERSION to set an extra version.  If
that's not passed and snabb is being built from a git checkout, "git
describe" will be used to obtain an extra version string.