lj_auditlog: Add "event" object
Log both "events" (new) and "memory" (already had this.)

The "event" and "memory" log messages are complementary. First we log
objects in memory, as detailed context, and then we log events, which
can refer to those objects.

For example, when a trace stops (i.e. is successfully recorded and
assembled), then we first log the GCtrace object in memory:

    {
        "type": "memory",
        "hint": "GCtrace",
        "address": 139883362546248,
        "data": <bin of size 120>
    }

and then later we log a "trace_stop" event that refers to this memory:

    {
        "type": "event",
        "event": "trace_stop",
        "GCtrace": 139883362546248
    }

So the "memory" messages are effectively providing context for later
"events".