Generate separate DWARF debug info
The build now produces a ".dwo" file containing DWARF debug
information. This contains a binary encoding of internal information
about RaptorJIT e.g. the layout of all its internal data structures.

The DWARF information is for use by debugging tools that need to
understand JIT internals.

The DWARF file is currently generated with gcc. I would prefer to use
clang, for the sake of consistency with the rest of the build, but I
seem to get _much_ more debug information from gcc so I stick with
that for now.

The nix build expression outputs the debug information as
lib/raptorjit.dwo