apps.socket.unix: Rewrote pull/push to avoid select()
Avoiding select() for two reasons. First, the way that we were passing
the fdset argument to select() - as a Lua table - is an NYI for the
JIT and forces the code to run interpreted. Second, we don't really
need select(), and can instead deal with failure or read() and write().