l2vpn: add extra_config to driver configuration
The configuration of a driver can have an optional element
extra_config, which must be a table and will be merged with the
regular config table.  This feature is intended to be used by
Lua-agnostic front-ends to the l2vpn program that create the
configuration from a higher layer of abstraction.