Addresses feedback on multiprocess changes
There were problems regarding regular (not on-a-stick) configuration not being
able to be configured with the new instance list. This adds a device leaf to the
external-interface per device container allowing to specify another card for that.

Moved the vlan grouping to per instance interface configuration allowing vlan tagging
to be physical interface specific.

Converted next-hop to a choice statement to allow for validation within yang. It also
adds a read only mac field when using IP to configure the next-hop to expose the resolved
mac address.

Added an mru leaf to external and internal interface containers to support specifying a
maximum recieve packet size.

Added versioning to the binding table. This includes a incrementing integer version number
and a date of last update.

Moves the statistics into a grouping statements and puts them on the interface as well
as the global counters.